Feat(Clientes): Cambios al imprimir afiliación
Nuevo formato al imprimir afilicación y envio al workmanager

+++ src/app/Services/Auth/auth-http-client-interceptor.service.ts
@@ -22,8 +22,13 @@ export class AuthHttpClientInterceptorService implements HttpInterceptor {
       });
     else
       authRequest = req.clone({});
+
+    const skipError = req.url.includes('skipErrorHandling=true');
     return next.handle(authRequest).pipe(
       catchError((err: any) => {
+        if (skipError) {
+          return throwError(() => err.error);
+        }
         if (err.status == 401) {
           localStorage.clear();
           window.location.reload();
